Fixed warnings about implicit type conversions.

+++ include/arr.h
@@ -588,7 +588,7 @@ public:
     {
         // Bounds check.
 
-        if ((uint)pos >= length())
+        if (pos >= length())
         {
             return NULL;
         }
